==> Translation model/Notebooks/01.ipynb <==
import os
import time
import numpy as np
import pandas as pd
import tensorflow as tf
import dill
import sys
sys.path.append("..")

from utils.helper import Tokenizer, preprocess_sentence

df1 = pd.read_csv("../Data/translation_train.csv",encoding='UTF-8')
df1.head(10)

df2 = pd.read_csv("../Data/ara.txt", sep='\t',encoding='UTF-8',header=None).drop(columns=[2]).rename(columns={0:"English",1:"Arabic"})
df2.head(10)

df = pd.concat([df1,df2],ignore_index=True)
df.head(10)

df.duplicated().sum()

df = df.drop_duplicates()

df.shape

df["English"] = df["English"].apply(lambda x: preprocess_sentence(x,"en"))
df["Arabic"] = df["Arabic"].apply(lambda x: preprocess_sentence(x,"ar"))

df.head(10)

if not os.path.exists("../Models/EN_Tokenizer.pkl"):
    EN = Tokenizer()
    EN.fit_on_texts(df["English"].values)
else:
    with open('../Models/EN_Tokenizer.pkl', 'rb') as file:
        EN = dill.load(file)

if not os.path.exists("../Models/AR_Tokenizer.pkl"):
    AR = Tokenizer()
    AR.fit_on_texts(df["Arabic"].values)
else: 
    with open('../Models/AR_Tokenizer.pkl', 'rb') as file:
        AR = dill.load(file)

df["English_seq"] = df["English"].apply(lambda x: EN.texts_to_sequences([x])[0])
EN.maxlen = df["English_seq"].apply(len).max()
df["Arabic_seq"] = df["Arabic"].apply(lambda x: AR.texts_to_sequences([x])[0])
AR.maxlen = df["Arabic_seq"].apply(len).max()
df.head()

df["English_seq"] = df["English_seq"].apply(lambda x: "_".join(np.array(x).astype(str)))
df["Arabic_seq"] = df["Arabic_seq"].apply(lambda x: "_".join(np.array(x).astype(str)))

with open('../Models/EN_Tokenizer.pkl', 'wb') as file:
    dill.dump(EN, file)

with open('../Models/AR_Tokenizer.pkl', 'wb') as file:
    dill.dump(AR, file)

df[["English_seq","Arabic_seq"]].to_csv("../Data/data.csv",index=False)
